refactor: place 관련 패키지를 turip 패키지로 통합 및 이동
