+++ android/app/src/main/java/com/on/turip/data/turip/TuripMapper.kt
@@ -3,11 +3,14 @@ package com.on.turip.data.turip
 import com.on.turip.data.turip.dto.TuripByPlaceResponse
 import com.on.turip.data.turip.dto.TuripCreationResponse
 import com.on.turip.data.turip.dto.TuripPatchRequest
+import com.on.turip.data.turip.dto.TuripPlacesResponse
 import com.on.turip.data.turip.dto.TuripPostRequest
 import com.on.turip.data.turip.dto.TuripResponse
 import com.on.turip.data.turip.dto.TuripsByPlaceResponse
 import com.on.turip.data.turip.dto.TuripsResponse
+import com.on.turip.domain.favorite.TuripPlace
 import com.on.turip.domain.folder.Turip
+import com.on.turip.domain.trip.Place
 
 fun TuripsResponse.toDomain(): List<Turip> = turipsResponse.map { it.toDomain() }
 
@@ -41,3 +44,21 @@ fun TuripByPlaceResponse.toDomain(): Turip =
         placeCount = 0,
         hasIncludePlace = isTuripPlace,
     )
+
+fun TuripPlacesResponse.toDomain(): List<TuripPlace> =
+    turipPlaceResponses.map { favoritePlace ->
+        TuripPlace(
+            id = favoritePlace.id,
+            order = favoritePlace.order,
+            place =
+                Place(
+                    placeId = favoritePlace.placeResponse.id,
+                    name = favoritePlace.placeResponse.name,
+                    url = favoritePlace.placeResponse.url,
+                    address = favoritePlace.placeResponse.address,
+                    latitude = favoritePlace.placeResponse.latitude,
+                    longitude = favoritePlace.placeResponse.longitude,
+                    category = favoritePlace.placeResponse.categories.map { it.name },
+                ),
+        )
+    }
