+++ android/app/src/main/java/com/on/turip/data/turip/datasource/DefaultTuripRemoteDataSource.kt
@@ -1,11 +1,11 @@
 package com.on.turip.data.turip.datasource
 
 import com.on.turip.core.result.TuripResult
-import com.on.turip.data.place.dto.TuripPlaceOrderRequest
-import com.on.turip.data.place.dto.TuripPlacesResponse
 import com.on.turip.data.result.safeApiCall
 import com.on.turip.data.turip.dto.TuripCreationResponse
 import com.on.turip.data.turip.dto.TuripPatchRequest
+import com.on.turip.data.turip.dto.TuripPlaceOrderRequest
+import com.on.turip.data.turip.dto.TuripPlacesResponse
 import com.on.turip.data.turip.dto.TuripPostRequest
 import com.on.turip.data.turip.dto.TuripsByPlaceResponse
 import com.on.turip.data.turip.dto.TuripsResponse
