+++ android/app/src/main/java/com/on/turip/data/turip/datasource/TuripRemoteDataSource.kt
@@ -1,10 +1,10 @@
 package com.on.turip.data.turip.datasource
 
 import com.on.turip.core.result.TuripResult
-import com.on.turip.data.place.dto.TuripPlaceOrderRequest
-import com.on.turip.data.place.dto.TuripPlacesResponse
 import com.on.turip.data.turip.dto.TuripCreationResponse
 import com.on.turip.data.turip.dto.TuripPatchRequest
+import com.on.turip.data.turip.dto.TuripPlaceOrderRequest
+import com.on.turip.data.turip.dto.TuripPlacesResponse
 import com.on.turip.data.turip.dto.TuripPostRequest
 import com.on.turip.data.turip.dto.TuripsByPlaceResponse
 import com.on.turip.data.turip.dto.TuripsResponse
