+++ android/app/src/main/java/com/on/turip/data/turip/dto/TuripPlaceOrderRequest.kt
@@ -1,4 +1,4 @@
-package com.on.turip.data.place.dto
+package com.on.turip.data.turip.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
